isolate_roots: Reuse computes signs of endpoints between iterations
For each iteration a lot of the endpoints of the intervals are the
same as in the previous iteration. Before the sign at the endpoint was
recomputed every time, not it reuses the computations for earlier
iterations. This reduces the number of calculations for the endpoints
by about a factor 4.